Convert from print() to Absl logging.
Existing print() statements are either eliminated or converted to logging statements.

Note, for the time being, the Absl logging library is configured to use STDOUT as its output stream.  This isn't a traditional logging configuration and might be worth reconsidering in the future.  However, for the time being, it provides the most work-alike behavior, for the transition to logging from print().

Incidental: Minor style fixes in entrez_lookup.py.