Specify types in Pandas dataframe columns, for accuracy and memory efficiency.
Segfaults on my test system were markedly reduced in frequency with the 60-fold (!!) reduction in memory usage for threshold dataframes (DIPLOID, HOMDEL, etc.).  Phenotype columns are now stored as strings, and Entrez IDs are now stored as 64-bit ints.

The Entrez change carried a change to the typespec of translate_gene_symbol_to_entrez_id(), which now returns an int instead of a string.  Furthermore, the test for that function is now updated to check for 'pandas.NA' (a legal 64-bit int in the Pandas type system) instead of 'None' (which Pandas regards as a type violation for 64-bit ints).

A small adjustment for init_data_tests applies the 'category' typing to test data there as well, so that dataframe equivalence comparisons remain valid.