Adjust the package structure so that "sys.path += ['./input_data_adapters']" is no longer needed.
This unfortunately comes with a new 'silly' arrangement in tcga_cnv_parser.py.  The new silliness seems better than the old silliness, because it no longer mucks about with system import paths.  However it's still something to try to eliminate, going forward.  A new detailed TODO comment offers some details and thoughts on the matter.

In general, the overall directory structure should perhaps be reconsidered altogether (e.g. Python seems to want a 'src' directory, whereas R wants 'R' and 'Script' directories).  That is left for another time; and for the time being I'm taking the elimination of "sys.path += ['./input_data_adapters']" as a win.